requirements.txt
certifi==2019.11.28
coverage==5.0.4
# Editable Git install with no remote (NLP==0.0.post0.dev1+g25d8c2f.dirty)
-e /media/data1/flo/NLP/src
nltk==3.4.5
numpy==1.18.2
pandas==1.0.3
PyScaffold==3.2.3
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
six==1.14.0
torch==1.4.0
torch-testing==0.0.2
